move ldap_config.py to configuration/, adjusts to docs
