+++ .gitignore
@@ -6,5 +6,4 @@ env_file_app
 env_file_postgres
 env_file_integrations
 venv/
-settings/ldap_config.py
 docker-compose-override.yml
